Move BCP controller to own folder
