+++ src/.editorconfig
@@ -24,7 +24,6 @@ resharper_blank_lines_around_single_line_auto_property = 1
 resharper_braces_for_ifelse = required
 resharper_csharp_insert_final_newline = true
 resharper_csharp_place_constructor_initializer_on_same_line = false
-resharper_csharp_trailing_comma_in_multiline_lists = true
 resharper_csharp_wrap_after_declaration_lpar = true
 resharper_csharp_wrap_after_invocation_lpar = true
 resharper_csharp_wrap_parameters_style = chop_if_long
