+++ src/STW.StubApi.Presentation/HealthChecks/HealthCheckOptionsBuilder.cs
