+++ src/Stylecop/stylecop.ruleset
@@ -14,9 +14,9 @@
         <Rule Id="SA1006" Action="Error"/>          <!-- Preprocessor keywords should not be preceded by space -->
         <Rule Id="SA1007" Action="Error"/>          <!-- Operator keyword should be followed by space -->
         <Rule Id="SA1008" Action="Error"/>          <!-- Opening parenthesis should be spaced correctly -->
-        <Rule Id="SA1009" Action="None"/>           <!-- Closing parenthesis should be spaced correctly -->
-        <Rule Id="SA1010" Action="None"/>           <!-- Opening square brackets should be spaced correctly -->
-        <Rule Id="SA1011" Action="None"/>           <!-- Closing square brackets should be spaced correctly -->
+        <Rule Id="SA1009" Action="Error"/>          <!-- Closing parenthesis should be spaced correctly -->
+        <Rule Id="SA1010" Action="Error"/>          <!-- Opening square brackets should be spaced correctly -->
+        <Rule Id="SA1011" Action="Error"/>          <!-- Closing square brackets should be spaced correctly -->
         <Rule Id="SA1012" Action="Error"/>          <!-- Opening braces should be spaced correctly -->
         <Rule Id="SA1013" Action="Error"/>          <!-- Closing braces should be spaced correctly -->
         <Rule Id="SA1014" Action="Error"/>          <!-- Opening generic brackets should be spaced correctly -->
@@ -71,7 +71,7 @@
         <Rule Id="SA1132" Action="Error"/>          <!-- Do not combine fields -->
         <Rule Id="SA1133" Action="Error"/>          <!-- Do not combine attributes -->
         <Rule Id="SA1134" Action="Error"/>          <!-- Attributes should not share line -->
-        <Rule Id="SA1135" Action="Error"/>          <!-- Using directives should be qualified -->
+        <Rule Id="SA1135" Action="None"/>           <!-- Using directives should be qualified -->
         <Rule Id="SA1136" Action="Error"/>          <!-- Enum values should be on separate lines -->
         <Rule Id="SA1137" Action="Error"/>          <!-- Elements should have the same indentation -->
         <Rule Id="SA1139" Action="Error"/>          <!-- Use literal suffix notation instead of casting -->
@@ -79,11 +79,11 @@
     </Rules>
 
     <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers.OrderingRules">
-        <Rule Id="SA1200" Action="None"/>           <!-- Using directives should be placed correctly -->
+        <Rule Id="SA1200" Action="Error"/>          <!-- Using directives should be placed correctly -->
         <Rule Id="SA1201" Action="Error"/>          <!-- Elements should appear in the correct order -->
         <Rule Id="SA1202" Action="Error"/>          <!-- Elements should be ordered by access -->
         <Rule Id="SA1203" Action="Error"/>          <!-- Constants should appear before fields -->
-        <Rule Id="SA1204" Action="None"/>           <!-- Static elements should appear before instance elements -->
+        <Rule Id="SA1204" Action="Error"/>          <!-- Static elements should appear before instance elements -->
         <Rule Id="SA1205" Action="Error"/>          <!-- Partial elements should declare access -->
         <Rule Id="SA1206" Action="Error"/>          <!-- Declaration keywords should follow order -->
         <Rule Id="SA1207" Action="Error"/>          <!-- Protected should come before internal -->
